Make everything use the database for global configuration
We permanently have two config providers now: One local, config-file
based one containing instance-specific information which is often tied
to a specific machine, and one global one which is kept in a MongoDB
database.
All Laniakea tools will connect to this database and try to find their
configuration there.

The lk-admin tool can be used to manipulate the database easily, and
also to set up the initial configuration.

This is all preliminary work, there can (and will be) bugs.