lighthouse: Refine communication with Spark workers
This patch fixes a bunch of issues and inconsistencies in the
communication of Lighthouse hubs with Spark worker daemons.
Most notably, two crashes are resolved, one due to invalid SQL, the
other due to the UUID parser being too pedantic about empty strings.

We also send less information to the workers now - we don't need to send
things that are not relevant for the job (and we also work around a Json
serialization issue that way).