Don't hardcode 3rd-party URLs anywhere in the code
If Laniakea automatically downloads stuff from a 3rd-party remote
location, it should now use the 3rd-party.json registry of stuff.
That makes it obvious which components might be downloaded and from
where. It also makes fixing URLs much easier in an installed instance,
should the URL ever change in future.