synchrotron: A "sync" may sometimes be a simple data import
This just changes terminology to be less confusing for users. "Import"
is much more clear than "synchronization", so we just call imports
imports if they are clearly just that.
If we (ever) perform more complex merges, they will be called merges and
the whole process is package synchronization. Simple, isn't it? ;-)