ariadne: Schedule arch:all builds for packages that don't build for indep-affinity
This handles cases where the permitted source architectures are "all
arm64" but our arch:all affinity is "amd64". prevuously, arch:all
packages wouldn't be built, now we enforce them to be built.

I am not happy with having the decision whether arch-indep packages are
built split between two Laniakea modules (Ariadne and Lighthouse), so
this may receive a future revision that folds the whole decision into
Ariadne, if that turns out to be the better approach.