ci: Rework automatic tests
Unfortunately we can't use the native Python support of GitHub Actions
as certain modules (like python-apt) have to come from the native
environment, which clashes with GitHub's Python environment.

With that considered, it also doesn't make sense to split the linting of
to its own workflow, so we will perform linting in a Debian Testing
environment now.
This patch also adds back a job to run our pytest tests and generally
restructures the CI a bit.