Use black + isort as our new coding style
While I do think that code styling written by actual humans is superior
to automated styling (as it can convey meaning that the automatic tool
doesn't know about), the amount of time fiddling with the style of code
and writing nitpick comments in PRs is IMHO no longer worth that
advantage.
So, let's switch to Black (but keep string literals the way they were,
for now).
Also sort all imports with isort, because doing that manually is just
annoying.