archive: Allow different src/bin version numbers in version memory
Source package and binary package versions may be different, and they
are in separate namespaces and need to be treated as such in the version
memory.
We add a src: prefix for source packages, which is a bit of a hack, but
should be fine for this application (as it gives us immediate uniqueness
constraints and we should not need to query e.g. the version memory of
all source packages ever).