Make binary packages not directly reference their overrides
A BinaryPackage was directly linked (in a one-to-one relationship) to a
PackageOverride, which was a schema design mistake, as this meant that
it was not only impossible to override data on a per-suite per-repo
basis, but also meant that the RepoSuiteSettings reference was pointing
to an inaccurate suite for a binary that is present in multiple suites.

This patchset fixes that mistake and also cleans up some code around
external package metadata handling that we can not share between the
archive and the repository reader anymore due to design differences.