Make Job entities reference their suites directly
Pretty much every job has the suite set in JSON, and this property is
also potentially useful for other job types as well. Therefore, we add
it as direct reference for improved efficiency.

In addition to that, Lighthouse will also now emit a packages'
repository to the worker, so we can dput to the right place.