core: Fix archive user updating their keys when using autoconfig
Previously a user's key fingerprint would be updated in the database but
not in the GPG keyring if the user already existed. This change will add
the key unconditionally if it is missing in the GPG keyring and ensure
it and the DB are in sync again.

This allows one key adding multiple subkeys.